+++ pkg/statemachine/client_hash_disseminator.go
@@ -10,6 +10,7 @@ import (
 	"bytes"
 	"container/list"
 	"fmt"
+	"github.com/hyperledger-labs/mirbft/pkg/logger"
 	"sort"
 
 	"github.com/hyperledger-labs/mirbft/pkg/pb/msgs"
@@ -119,7 +120,7 @@ import (
 // replicatingMap -- all requests which are currently being fetched
 
 type clientHashDisseminator struct {
-	logger      Logger
+	logger      logger.Logger
 	myConfig    *state.EventInitialParameters
 	nodeBuffers *nodeBuffers
 
@@ -131,7 +132,7 @@ type clientHashDisseminator struct {
 	clientTracker    *clientTracker
 }
 
-func newClientHashDisseminator(nodeBuffers *nodeBuffers, myConfig *state.EventInitialParameters, logger Logger, clientTracker *clientTracker) *clientHashDisseminator {
+func newClientHashDisseminator(nodeBuffers *nodeBuffers, myConfig *state.EventInitialParameters, logger logger.Logger, clientTracker *clientTracker) *clientHashDisseminator {
 	return &clientHashDisseminator{
 		logger:        logger,
 		myConfig:      myConfig,
@@ -669,7 +670,7 @@ func (cr *clientRequest) fetch() *ActionList {
 
 type client struct {
 	myConfig      *state.EventInitialParameters
-	logger        Logger
+	logger        logger.Logger
 	networkConfig *msgs.NetworkState_Config
 	clientState   *msgs.NetworkState_Client
 	clientTracker *clientTracker
@@ -681,7 +682,7 @@ type client struct {
 	reqNoMap  map[uint64]*list.Element
 }
 
-func newClient(myConfig *state.EventInitialParameters, logger Logger, tracker *clientTracker) *client {
+func newClient(myConfig *state.EventInitialParameters, logger logger.Logger, tracker *clientTracker) *client {
 	return &client{
 		myConfig:      myConfig,
 		logger:        logger,
@@ -737,7 +738,7 @@ func (c *client) reinitialize(seqNo uint64, networkConfig *msgs.NetworkState_Con
 
 	c.advanceReady()
 
-	c.logger.Log(LevelDebug, "reinitialized client", "client_id", c.clientState.Id, "low_watermark", c.clientState.LowWatermark, "high_watermark", c.highWatermark, "next_ready_mark", c.nextReadyMark, "next_ack_mark", c.nextAckMark)
+	c.logger.Log(logger.LevelDebug, "reinitialized client", "client_id", c.clientState.Id, "low_watermark", c.clientState.LowWatermark, "high_watermark", c.highWatermark, "next_ready_mark", c.nextReadyMark, "next_ack_mark", c.nextAckMark)
 
 	return actions
 }
@@ -798,7 +799,7 @@ func (c *client) allocate(seqNo uint64, state *msgs.NetworkState_Client, reconfi
 	c.highWatermark = newHighWatermark
 	c.advanceReady()
 
-	c.logger.Log(LevelDebug, "allocated new reqs for client", "client_id", c.clientState.Id, "low_watermark", c.clientState.LowWatermark, "high_watermark", c.highWatermark, "next_ready_mark", c.nextReadyMark)
+	c.logger.Log(logger.LevelDebug, "allocated new reqs for client", "client_id", c.clientState.Id, "low_watermark", c.clientState.LowWatermark, "high_watermark", c.highWatermark, "next_ready_mark", c.nextReadyMark)
 
 	return actions
 }
