--- serializer.go
@@ -1,20 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package mirbft
-
-import (
-	"github.com/hyperledger-labs/mirbft/pkg/statemachine"
-)
-
-// TODO, add assertion in tests that log levels match
-type LogAdapter struct {
-	Logger
-}
-
-func (la LogAdapter) Log(level statemachine.LogLevel, msg string, args ...interface{}) {
-	la.Logger.Log(LogLevel(level), msg, args...)
-}
